+++ src/main/java/com/flytrap/rssreader/api/account/business/service/AccountService.java
@@ -3,8 +3,8 @@ package com.flytrap.rssreader.api.account.business.service;
 import com.flytrap.rssreader.api.account.domain.Account;
 import com.flytrap.rssreader.api.account.domain.AccountId;
 import com.flytrap.rssreader.api.account.domain.AccountName;
-import com.flytrap.rssreader.api.account.infrastructure.repository.AccountCommand;
-import com.flytrap.rssreader.api.account.infrastructure.repository.AccountQuery;
+import com.flytrap.rssreader.api.account.infrastructure.implement.AccountCommand;
+import com.flytrap.rssreader.api.account.infrastructure.implement.AccountQuery;
 import com.flytrap.rssreader.global.exception.domain.NoSuchDomainException;
 import com.flytrap.rssreader.api.auth.infrastructure.external.dto.UserResource;
 import jakarta.transaction.Transactional;
