+++ src/main/java/com/flytrap/rssreader/api/account/infrastructure/implement/AccountCommand.java
@@ -1,7 +1,8 @@
-package com.flytrap.rssreader.api.account.infrastructure.repository;
+package com.flytrap.rssreader.api.account.infrastructure.implement;
 
 import com.flytrap.rssreader.api.account.domain.Account;
 import com.flytrap.rssreader.api.account.infrastructure.entity.AccountEntity;
+import com.flytrap.rssreader.api.account.infrastructure.repository.AccountJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
