+++ src/main/java/com/flytrap/rssreader/api/account/infrastructure/implement/AccountQuery.java
@@ -1,9 +1,10 @@
-package com.flytrap.rssreader.api.account.infrastructure.repository;
+package com.flytrap.rssreader.api.account.infrastructure.implement;
 
 import com.flytrap.rssreader.api.account.domain.Account;
 import com.flytrap.rssreader.api.account.domain.AccountId;
 import com.flytrap.rssreader.api.account.domain.AccountName;
 import com.flytrap.rssreader.api.account.infrastructure.entity.AccountEntity;
+import com.flytrap.rssreader.api.account.infrastructure.repository.AccountJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
